# api_server/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: info@payloadops.com
    name: Payload Ops Team
  description: "The Payload Ops API streamlines management of AI prompts, projects,\
    \ organizations, teams, and users through conventional HTTP requests. This platform\
    \ enables efficient automation and control of resources, providing robust tools\
    \ for developers to manage settings, memberships, and activities seamlessly."
  title: Payload Ops API
  version: "1.0"
servers:
- description: Production server (uses live data)
  url: https://api.payloadops.com/v1
- description: Sandbox server (uses test data)
  url: https://sandbox-api.payloadops.com/v1
security:
- ApiKeyAuth: []
- BearerAuth: []
paths:
...
  /projects/{projectId}/keys:
    get:
      description: |
        Retrieves a list of all API keys associated with a specific project. This endpoint is essential for managing and reviewing the access provided to different services and users. It offers a comprehensive view of all keys, their scopes, and their statuses, which is critical for effective access management and security audits.

        **Features**:
        - **Comprehensive Overview**: Provides a complete listing of all API keys, facilitating audits and security reviews.
        - **Access Transparency**: Ensures that administrators can view and monitor the scopes and usage of each API key, aiding in proactive security practices.

        **Use Cases**:
        - Auditing API keys to ensure that they are up-to-date and have appropriate permissions.
        - Quickly identifying and managing keys that may need adjustments or revocation.
      operationId: listApiKeys
      parameters:
      - description: The unique identifier of the project whose API keys are to be
          listed.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ApiKey'
                type: array
          description: Successfully retrieved a list of API keys for the project.
        "404":
          description: "The specified project was not found, indicating an invalid\
            \ project ID."
        "500":
          description: "A server error occurred, preventing the retrieval of API keys."
      security:
      - BearerAuth: []
      summary: List all API keys for a project
      tags:
      - API Keys
    post:
      description: |
        Generates a new API key for a specific project with designated scopes. This operation is crucial for creating secure, scoped access to the project's resources. It allows administrators to define specific permissions, ensuring that API keys can only access appropriate endpoints as per their operational requirements.

        **Features**:
        - **Scoped Access**: Tailors access permissions to specific needs, enhancing security by limiting API key capabilities to necessary actions.
        - **Secure Management**: Generates keys in a secure manner, ensuring that they are only known to the requester and stored securely.

        **Use Cases**:
        - Provisioning a new service or system that requires access to the project.
        - Creating temporary access for automation scripts or third-party integrations.

        **Guidelines**:
        - Ensure to specify the necessary scopes in the request body to avoid overly permissive access.
        - Manage the lifecycle of API keys responsibly, revoking them when no longer needed.
      operationId: generateApiKey
      parameters:
      - description: The identifier of the project for which the API key is being
          generated.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyInput'
        description: Contains the scope definitions for the new API key.
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
          description: New API key generated successfully. The key details are included
            in the response.
        "400":
          description: "Invalid request, such as missing required fields or invalid\
            \ scope specifications."
        "404":
          description: The specified project was not found.
        "500":
          description: "A server error occurred, preventing the generation of the\
            \ API key."
      security:
      - BearerAuth: []
      summary: Generate a new API key with specific scopes for a project
      tags:
      - API Keys
  /projects/{projectId}/keys/{keyId}:
    delete:
      description: |
        Deletes a specific API key from a project, permanently removing its access to the project resources. This operation is essential for maintaining security when an API key is no longer needed or if its permissions need to be revoked urgently.

        **Key Features**:
        - **Secure Cleanup**: Ensures that unused or compromised API keys are removed to prevent unauthorized access.
        - **Access Revocation**: Provides a method to immediately invalidate an API key when necessary.

        **Use Cases**:
        - Removing access for team members who no longer require it, such as after project completion or role changes.
        - Revoking keys that may have been exposed or compromised to maintain security integrity.
      operationId: deleteApiKey
      parameters:
      - description: The unique identifier of the project from which the API key will
          be deleted.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the API key to be deleted.
        explode: false
        in: path
        name: keyId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: "The API key was deleted successfully, no content returned."
        "404":
          description: Either the API key or the project was not found.
        "500":
          description: "A server error occurred, preventing the deletion of the API\
            \ key."
      security:
      - BearerAuth: []
      summary: Delete a specific API key
      tags:
      - API Keys
    get:
      description: |
        Retrieves detailed information about a specific API key associated with a project. This operation is crucial for managing access to project resources, allowing administrators to review the permissions and details of individual API keys.

        **Key Features**:
        - **Security Management**: Provides the ability to monitor and manage API access through detailed key information.
        - **Access Review**: Enables project managers to verify the scope and validity of an API key, ensuring that it aligns with current security policies.

        **Use Cases**:
        - Administrators auditing access permissions for compliance and security best practices.
        - Developers checking API key details to troubleshoot access issues or to ensure proper configuration.
      operationId: getApiKey
      parameters:
      - description: The unique identifier of the project associated with the API
          key.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the API key to retrieve.
        explode: false
        in: path
        name: keyId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
          description: Detailed information about the API key retrieved successfully.
        "404":
          description: Either the specified API key or the project was not found.
        "500":
          description: "A server error occurred, preventing the retrieval of the API\
            \ key."
      security:
      - BearerAuth: []
      summary: Retrieve a specific API key
      tags:
      - API Keys
    put:
      description: |
        Updates the scopes associated with a specific API key. This operation allows administrators to modify access levels and permissions for API keys, adapting to changing security needs or project requirements.

        **Key Features**:
        - **Scope Management**: Adjusts the access boundaries defined by the API key scopes, ensuring they are kept up-to-date with project roles and permissions.
        - **Flexible Access Control**: Offers the ability to dynamically change API key permissions without needing to recreate keys.

        **Use Cases**:
        - Modifying API key permissions in response to role changes within a project team.
        - Updating key scopes to restrict or expand access as part of security policy adjustments.
      operationId: updateApiKey
      parameters:
      - description: The unique identifier of the project for which the API key is
          managed.
        explode: false
        in: path
        name: projectId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the API key to be updated.
        explode: false
        in: path
        name: keyId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyInput'
        description: JSON payload containing the new scopes for the API key.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
          description: The API key's scopes were updated successfully.
        "400":
          description: "Invalid input, such as unspecified or unsupported scopes."
        "404":
          description: Either the API key or the project was not found.
        "500":
          description: "A server error occurred, preventing the update of the API\
            \ key."
      security:
      - BearerAuth: []
      summary: Update an API key's scopes
      tags:
      - API Keys
...
  /prompts/{promptId}/branches/{branchId}/commits:
    get:
      description: |
        Retrieves a list of all commit histories associated with a specific branch of a prompt. This endpoint is crucial for understanding the evolution and lifecycle of branch content, allowing users to access and review all historical and current iterations of a branch.

        **Features**:
        - **Commit Tracking**: Provides detailed information about each commit, including commit messages, authors, and timestamps, which helps in tracking the changes made over time.
        - **Historical Analysis**: Facilitates the analysis of how a branch has evolved, enabling users to see the progression of content and identify key milestones.
        - **Auditing and Compliance**: Maintains a clear record of all changes made to a branch, supporting auditing and compliance requirements.

        **Use Cases**:
        - Developers reviewing commit histories to understand the development process of a branch.
        - Content managers analyzing changes for quality control and consistency.
        - Compliance officers auditing changes for regulatory adherence.
      operationId: listBranchCommits
      parameters:
      - description: The unique identifier of the prompt whose branch commits are
          being listed.
        explode: false
        in: path
        name: promptId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the branch whose commits are being listed.
        explode: false
        in: path
        name: branchId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Commit'
                type: array
          description: Successfully retrieved a list of all commits associated with
            the branch.
        "404":
          description: No branch found matching the specified identifiers.
        "500":
          description: "A server error occurred, preventing the retrieval of commit\
            \ information."
      summary: List all commits of a specific branch
      tags:
      - Commits
    post:
      description: |
        Creates a new commit for a specific branch of a prompt. This endpoint allows users to record changes made to the branch, with a commit message detailing the updates. It supports workflows where changes need to be tracked and managed systematically.

        **Features**:
        - **Change Documentation**: Captures detailed descriptions of changes made in the commit message, providing context for future reference.
        - **Version Control**: Ensures systematic tracking of all changes, enabling easy rollback to previous states if needed.
        - **Collaborative Development**: Supports collaborative environments by recording contributions from multiple users.

        **Use Cases**:
        - Developers committing new changes or features to a branch.
        - Content managers updating prompt content and documenting the changes.
        - Teams collaborating on a project and needing a clear record of each member's contributions.
      operationId: createBranchCommit
      parameters:
      - description: The unique identifier of the prompt whose branch is being committed
          to.
        explode: false
        in: path
        name: promptId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the branch for which a new commit is
          being created.
        explode: false
        in: path
        name: branchId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitInput'
        description: "Contains the necessary information for creating a new commit,\
          \ such as commit message and changes."
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commit'
          description: Commit created successfully and is now part of the branch history.
        "400":
          description: "Invalid input, such as missing required fields or malformed\
            \ data."
        "404":
          description: "The specified branch was not found, and thus the commit could\
            \ not be created."
        "500":
          description: "A server error occurred, preventing the creation of a new\
            \ commit."
      summary: Create a new commit for a branch
      tags:
      - Commits
  /prompts/{promptId}/branches/{branchId}/commits/{commitId}:
    get:
      description: |
        Retrieves detailed information about a specific commit of a branch, or the most recent commit if 'latest' is specified as the commit ID. This endpoint is crucial for accessing historical data of branch commits for review or rollback purposes.

        **Features**:
        - **Historical Tracking**: Provides access to the specific details of a commit, including changes made, author, and timestamp.
        - **Latest Commit Access**: Supports retrieving the most recent commit using the 'latest' keyword, streamlining access to the current state of the branch.
        - **Detailed Insights**: Includes comprehensive commit information, aiding in understanding the evolution and decision-making process.

        **Use Cases**:
        - Retrieving past commits for audit or comparison purposes.
        - Accessing the latest commit to continue development or review current changes.
        - Integrating commit data into other systems for enhanced content lifecycle management.
      operationId: getBranchCommit
      parameters:
      - description: The unique identifier of the prompt.
        explode: false
        in: path
        name: promptId
        required: true
        schema:
          type: string
        style: simple
      - description: The unique identifier of the branch.
        explode: false
        in: path
        name: branchId
        required: true
        schema:
          type: string
        style: simple
      - description: "The ID of the commit to retrieve, or 'latest' to obtain the\
          \ most recent commit."
        example: latest
        explode: false
        in: path
        name: commitId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Commit'
          description: Successfully retrieved detailed information about the specified
            commit of the branch.
        "404":
          description: No branch or commit found matching the specified identifiers.
        "500":
          description: "A server error occurred, preventing the retrieval of the branch\
            \ commit."
      summary: Retrieve a specific commit or the latest commit of a branch
      tags:
      - Commits
components:
  schemas:
...
    ApiKey:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        id: id
        scopes:
        - scopes
        - scopes
        projectId: projectId
        key: key
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Unique identifier for the API key
          type: string
        key:
          description: The API key token
          type: string
        scopes:
          description: List of scopes or permissions granted by this API key
          items:
            type: string
          type: array
        projectId:
          description: The project ID this API key belongs to
          type: string
        createdAt:
          description: Timestamp when the API key was created
          format: date-time
          type: string
        updatedAt:
          description: Timestamp when the API key was last updated
          format: date-time
          type: string
      type: object
    ApiKeyInput:
      example:
        scopes:
        - scopes
        - scopes
      properties:
        scopes:
          description: List of scopes or permissions to be granted to this API key
          items:
            type: string
          type: array
      required:
      - scopes
      type: object
...
    Commit:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        id: id
        content: content
      properties:
        id:
          description: Unique identifier for the commit of the branch
          type: string
        content:
          description: Content of the prompt in this commit
          type: string
        createdAt:
          description: Timestamp when this commit was created
          format: date-time
          type: string
      type: object
    CommitInput:
      example:
        content: content
      properties:
        content:
          description: Content of the prompt in this commit
          type: string
      required:
      - content
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: |
        API Key authentication provides a straightforward method for clients to authenticate requests. This scheme expects an API key to be supplied in the headers of each request. Clients must include the API key in the 'X-API-KEY' header field.

        **How to use**:
        - Obtain an API key from the API provider.
        - Include the API key in the header of each request to the API as follows:
          ```
          X-API-KEY: {api_key}
          ```
        - The API validates the key against its database, and if the key is valid, allows access to the requested resources.

        **Advantages**:
        - Simple and easy to implement.
        - Suitable for accessing public APIs where security requirements are not excessively strict.

        **Limitations**:
        - Less secure than more complex authentication methods, such as tokens, as API keys can be easily exposed in transit or storage.
        - API keys do not expire by default and revoking them requires direct intervention.
      in: header
      name: X-API-KEY
      type: apiKey
    BearerAuth:
      bearerFormat: JWT
      description: |
        Bearer Token authentication uses JSON Web Tokens (JWT) to verify the identity of the requester. This scheme is included in the Authorization header of each HTTP request using the Bearer format.

        **How to use**:
        - Obtain a JWT from the authentication server. Typically, this involves an initial request with credentials (username/password) to receive the token.
        - Include the token in the Authorization header of your HTTP requests as follows:
          ```
          Authorization: Bearer {token}
          ```
        - The server decodes the JWT to verify its validity and authorizes the request based on the token's payload and signature.

        **Advantages**:
        - Offers more security than API key authentication as the token content is protected by cryptographic algorithms.
        - Tokens can be designed to expire after a certain period, enhancing security by limiting the duration of their validity.

        **Limitations**:
...