# api_server/go/model_commit.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Payload Ops API
 *
 * The Payload Ops API streamlines management of AI prompts, projects, organizations, teams, and users through conventional HTTP requests. This platform enables efficient automation and control of resources, providing robust tools for developers to manage settings, memberships, and activities seamlessly.
 *
 * API version: 1.0
 * Contact: info@payloadops.com
 */

package openapi


import (
	"time"
)



type Commit struct {

	// Unique identifier for the commit of the branch
	Id string `json:"id,omitempty"`

	// Content of the prompt in this commit
	Content string `json:"content,omitempty"`

	// Timestamp when this commit was created
	CreatedAt time.Time `json:"createdAt,omitempty"`
}

// AssertCommitRequired checks if the required fields are not zero-ed
func AssertCommitRequired(obj Commit) error {
	return nil
}

// AssertCommitConstraints checks if the values respects the defined constraints
func AssertCommitConstraints(obj Commit) error {
	return nil
}